refactor(mitm-addon): split body utility modules
