+++ crates/runner/mitm-addon/tests/test_response_streaming.py
@@ -5,13 +5,13 @@ import gzip
 import pytest
 from mitmproxy.test import tutils
 
-import body_utils
 import mitm_addon
 import response_streaming
+from body_limits import LARGE_RESPONSE_DECOMPRESS_LIMIT
 from tests.flow_helpers import header_map, response_stream
 from tests.x_flow_helpers import make_x_response_flow
 
-_OVERSIZED_NDJSON_LINE_BYTES = body_utils.LARGE_RESPONSE_DECOMPRESS_LIMIT + 1024
+_OVERSIZED_NDJSON_LINE_BYTES = LARGE_RESPONSE_DECOMPRESS_LIMIT + 1024
 
 
 class TestNdjsonExtractor:
